Generalizing the Palette code to use an arbitrary screen filter which can support any modification through render.InPlace()
